# Conflicts:
#	app/Http/Controllers/Auth/RegisteredUserController.php
#	resources/views/frontend/prime_x/withdraw/now.blade.php

+++ app/Http/Controllers/Backend/RebateRuleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreRebateRuleRequest;
+use App\Models\MultiLevel;
 use App\Models\RebateRule;
 use App\Models\SymbolGroup;
 use App\Services\RebateRuleService;
@@ -20,12 +21,12 @@ class RebateRuleController extends Controller
     public function index(Request $request)
     {
         if ($request->ajax()) {
-            $data = RebateRule::with('groups')->latest('updated_at')->get();
-    
+            $data = RebateRule::with('symbolGroups')->get();
+
             return Datatables::of($data)
                 ->addIndexColumn()
-                ->addColumn('groups', function($row) {
-                    return $row->groups->pluck('symbol_group')->implode(', ');
+                ->addColumn('symbolGroups', function($row) {
+                    return $row->symbolGroups->pluck('title')->implode(', ');
                 })
                 ->addColumn('status', 'backend.rebate_rules.include.__status')
                 ->addColumn('action', 'backend.rebate_rules.include.__action')
@@ -37,7 +38,7 @@ class RebateRuleController extends Controller
 
     public function create()
     {
-        $symbolGroups = SymbolGroup::pluck('symbol_group','id')->toArray();
+        $symbolGroups = SymbolGroup::pluck('title','id')->toArray();
         return response()->json(['symbolGroups'=>$symbolGroups]);
     }
     public function store(StoreRebateRuleRequest $request)
@@ -50,16 +51,13 @@ class RebateRuleController extends Controller
             return response()->json(['success' => false, 'message' => $e->getMessage()], 500);
         }
     }
-   
+
     public function edit($id)
     {
-        $rebateRule = RebateRule::with('groups')->find($id);
+        $rebateRule = RebateRule::with('symbolGroups')->find($id);
         $allSymbolGroups = SymbolGroup::all();
-    
-        return response()->json([
-            'rebateRule' => $rebateRule,
-            'allSymbolGroups' => $allSymbolGroups
-        ]);
+
+        return view('backend.rebate_rules.include.__edit_form', compact('rebateRule', 'allSymbolGroups'));
     }
     public function show(RebateRule $rebateRule)
     {
@@ -71,10 +69,14 @@ class RebateRuleController extends Controller
         notify()->success(__('Rebate Rule updated successfully.'));
         return redirect()->route('admin.rebate-rules.index');
     }
-   
+
 
     public function destroy(RebateRule $rebateRule)
     {
+        if($rebateRule->multiLevels()->count() > 0) {
+            notify()->error(__('Sorry,Cannot delete this rebate rule because it is still associated with multi-levels. Please detach the levels first'));
+            return redirect()->back();
+        }
         $this->rebateRuleService->delete($rebateRule);
         notify()->success(__('Rebate Rule deleted successfully.'));
         return redirect()->route('admin.rebate-rules.index');
